move testCreateStore() from mwp-test-utils to mwp-store/src/util/testUtils.js
