+++ packages/mwp-store/src/server/index.test.js
@@ -1,4 +1,4 @@
-import { testCreateStore } from 'mwp-test-utils';
+import { testCreateStore } from '../util/testUtils';
 import { getServerCreateStore } from './';
 
 const MOCK_ROUTES = {};
