+++ packages/mwp-test-utils/src/index.js
@@ -73,21 +73,3 @@ export async function getServer() {
 
 	return server;
 }
-
-const IDENTITY_REDUCER = state => state;
-export function testCreateStore(createStoreFn) {
-	it('creates a store with store functions', () => {
-		const basicStore = createStoreFn(IDENTITY_REDUCER);
-		expect(basicStore.getState).toEqual(jasmine.any(Function));
-		expect(basicStore.dispatch).toEqual(jasmine.any(Function));
-	});
-	it('creates a store with supplied initialState', done => {
-		const initialState = { foo: 'bar' };
-		const basicStore = createStoreFn(IDENTITY_REDUCER, initialState);
-		basicStore.subscribe(() => {
-			expect(basicStore.getState()).toEqual(initialState);
-			done();
-		});
-		basicStore.dispatch({ type: 'dummy' });
-	});
-}
